--- a/spec/support/components/work_packages/settings_menu.rb
+++ b/spec/support/components/work_packages/settings_menu.rb
@@ -34,17 +34,20 @@ module Components
       include RSpec::Matchers
 
       def open_and_save_query(name)
-        open!
-        find("#{selector} .menu-item", text: 'Save', match: :prefer_exact).click
-        page.within('.spot-modal') do
-          find('#save-query-name').set name
-          click_on 'Save'
-        end
+        open_and_choose('Save')
+        within_modal_fill_in_and_save(name:)
+      end
+
+      def open_and_save_query_as(name)
+        open_and_choose('Save as ...')
+        within_modal_fill_in_and_save(name:)
       end
 
       def open_and_choose(name)
-        open!
-        choose(name)
+        retry_block do
+          open!
+          choose(name)
+        end
       end
 
       def open!
@@ -57,11 +60,11 @@ module Components
       end
 
       def expect_closed
-        expect(page).to have_no_selector(selector)
+        expect(page).not_to have_selector(selector)
       end
 
       def choose(target)
-        find("#{selector} .menu-item", text: target).click
+        find("#{selector} .menu-item", text: target, match: :prefer_exact).click
       end
 
       def expect_options(options)
@@ -75,6 +78,22 @@ module Components
 
       def selector
         '#settingsDropdown'
+      end
+
+      def within_modal_fill_in_and_save(name:)
+        modal.within_modal do
+          fill_in 'save-query-name', with: name
+          click_button 'Save'
+        end
+        wait_for_save_completion
+      end
+
+      def wait_for_save_completion
+        modal.expect_closed
+      end
+
+      def modal
+        Components::Common::Modal.new
       end
     end
   end