--- a/spec/support/shared/selenium_workarounds.rb
+++ b/spec/support/shared/selenium_workarounds.rb
@@ -31,12 +31,18 @@
 # https://bugs.chromium.org/p/chromedriver/issues/detail?id=1771
 module SeleniumWorkarounds
   def ensure_value_is_input_correctly(input, value:)
-    # Wait a bit to insert the value
-    sleep(0.5)
-    input.set value
-    sleep(0.5)
+    correctly_set = false
+    # Wait longer and longer to set the value, until it is set correctly.
+    # The bug may be fixed by now...
+    [0, 0.5, 1].each do |waiting_time|
+      sleep(waiting_time)
+      input.set value
+      sleep(waiting_time)
 
-    found_value = input.value
-    raise "Found value #{found_value}, but expected #{value}." unless found_value == value
+      correctly_set = (input.value == value)
+      break if correctly_set
+    end
+
+    raise "Found value #{found_value}, but expected #{value}." unless correctly_set
   end
 end