… spec/support/matchers/raise_if_found.rb