`spec/features/work_packages/table/queries/filter_spec.rb` was particularly slow.

Selectors like `has_selector?` wait till the end of the timeout (5 secs) if the element is not present.
When the page is known to be fully loaded, adding `wait: false` saves waiting 5 seconds for nothing.

`spec/features/work_packages/table/queries/filter_spec.rb` spec run time:
- before: 6 minutes 38 seconds
- after 1st commit: 2 minutes 41 seconds
- after 2nd commit: 1 minute 16 seconds